Agregar HTML a DataTable lado del servidor

+++ cms/app/Http/Controllers/AdministradoresController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Administradores;
 use App\Blog;
+use Illuminate\Support\Facades\Hash;
 
 class AdministradoresController extends Controller
 {
@@ -23,6 +24,24 @@ class AdministradoresController extends Controller
         if(request()->ajax()){
 
             return datatables()->of(Administradores::all())
+            -> addColumn('acciones', function($data){
+
+                $acciones = '<div class="btn-group">
+
+                                <a href="'.url()->current().'/'.$data->id.'" class="btn btn-warning btn-sm">
+                                    <i class="fas fa-pencil-alt text-white"></i>
+                                </a>
+
+                                <button class="btn btn-danger btn-sm eliminarRegistro" action="'.url()->current().'/'.$data->id.'" method="DELETE" pagina="administradores" token="'.csrf_token().'">
+                                    <i class="fas fa-trash-alt"></i>
+                                </button>
+
+                            </div>';
+
+                return $acciones;
+
+            })
+            -> rawColumns(['acciones'])
             -> make(true);
 
         }
@@ -86,7 +105,7 @@ class AdministradoresController extends Controller
                 
                 $validarPassword = \Validator::make($password,[
 
-                    "passsword" => "required|regex:/^[0-9a-zA-Z]+$/i"
+                    "password" => "required|regex:/^[0-9a-zA-Z]+$/i"
 
                 ]);
 
@@ -182,7 +201,11 @@ class AdministradoresController extends Controller
 
         if(!empty($validar) && $id != 1){
 
-            unlink($validar[0]["foto"]);
+            if(!empty($validar[0]["foto"])){
+
+                unlink($validar[0]["foto"]);    
+
+            }
 
             $administradores = Administradores::where("id", $validar[0]["id"])->delete();
